+++ database/migrations/2025_01_29_101102_create_break_fast_report_products_table.php
@@ -14,8 +14,8 @@ return new class extends Migration {
             $table->id();
             $table->foreignIdFor(\App\BreakFast\BreakFastProduct::class)->constrained()->cascadeOnDelete()->cascadeOnUpdate();
             $table->foreignIdFor(\App\BreakFast\BreakFastReport::class)->constrained()->cascadeOnDelete()->cascadeOnUpdate();
-            $table->float('daily_amount', 10, 6)->nullable();
-            $table->float('price', 10, 6)->nullable();
+            $table->double('daily_amount', 20, 6)->nullable();
+            $table->double('price', 20, 6)->nullable();
             $table->timestamps();
         });
     }
